requirements.txt
anyio==4.11.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
cloudevents==1.11.0
deprecation==2.1.0
Flask==3.1.2
functions-framework==3.9.2
google-api-core==2.25.1
google-auth==2.41.0
google-cloud-core==2.4.3
google-cloud-pubsub==2.31.1
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.75.1
grpcio-status==1.75.1
gunicorn==23.0.0
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
-e git+ssh://git@github.com/7893/ns.git@bad78f280e3a40a78fa4cde3a478cf9cc3026784#egg=ns_packages&subdirectory=packages
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
packaging==25.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
starlette==0.48.0
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
uvicorn-worker==0.4.0
watchdog==6.0.0
Werkzeug==3.1.3
zipp==3.23.0
